Update try/catch to if block because the exception is IOException instead of FileNotFoundException

+++ server/src/com/mirth/connect/util/messagewriter/MessageWriterArchive.java
@@ -11,6 +11,7 @@ package com.mirth.connect.util.messagewriter;
 
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.io.IOException;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
@@ -74,19 +75,17 @@ public class MessageWriterArchive implements MessageWriter {
         if (messagesWritten) {
             try {
                 File tempFile = new File(archiveFile.getParent() + IOUtils.DIR_SEPARATOR + "." + archiveFile.getName());
-
-                try {
+                
+                if (tempFile.exists()) {
                     FileUtils.forceDelete(tempFile);
-                } catch (FileNotFoundException e) {
                 }
 
                 ArchiveUtils.createArchive(rootFolder, tempFile, archiver, compressor, password, encryptionType);
 
-                try {
+                if (archiveFile.exists()) { 
                     FileUtils.forceDelete(archiveFile);
-                } catch (FileNotFoundException e) {
                 }
-
+                
                 FileUtils.moveFile(tempFile, archiveFile);
             } catch (Exception e) {
                 throw new MessageWriterException(e);
